отдельный компо для редактирования заявки
